requirements.txt
fastapi
uvicorn
python-multipart
openai
agents
python-dotenv
aiofiles
requests
httpx
firebase-admin
google-generativeai


